Separated hero selection from hero display
