+++ src/components/GearPreview/index.js
@@ -1,0 +1,3 @@
+import GearPreview from './GearPreview';
+
+export default GearPreview;
