+++ src/components/HeroView/HeroView.js
@@ -1,3 +1,4 @@
+import Paper from '@material-ui/core/Paper';
 import withStyles from '@material-ui/core/styles/withStyles';
 import Typography from '@material-ui/core/Typography';
 import classNames from 'classnames';
@@ -8,33 +9,31 @@ import HeroStats from '../HeroStats';
 import Zodiac from '../Zodiac';
 
 const style = theme => ({
-	root: {},
+	root: {
+		padding: theme.spacing.unit,
+	},
 	title: {
 		display: 'inline',
 		marginRight: theme.spacing.unit,
 	},
 	head: {
-		padding: theme.spacing.unit * 3,
-		paddingBottom: theme.spacing.unit,
+		padding: theme.spacing.unit * 2,
+		paddingBottom: 0,
 	},
 });
 
 const HeroView = ({classes, className, hero, ...props}) => (
-	<div className={classNames(classes.root, className)} {...props}>
-		{hero && (
-			<React.Fragment>
-				<div className={classes.head}>
-					<Typography className={classes.title} variant="h4" component="span">
-						{hero.name}
-					</Typography>
-					<HeroStars awakened={hero.rarity} total={hero.rarity} />
-					<Element variant={hero.element} />
-					<Zodiac variant={hero.zodiac} />
-				</div>
-				<HeroStats stats={hero.stats} />
-			</React.Fragment>
-		)}
-	</div>
+	<Paper className={classNames(classes.root, className)} {...props}>
+		<div className={classes.head}>
+			<Typography className={classes.title} variant="h4" component="span">
+				{hero.name}
+			</Typography>
+			<HeroStars awakened={hero.rarity} total={6} />
+			<Element variant={hero.element} />
+			<Zodiac variant={hero.zodiac} />
+		</div>
+		<HeroStats stats={hero.stats} />
+	</Paper>
 );
 
 export default withStyles(style)(HeroView);
